Remove methods on `ClientBulder` that don't serve a purpose (#2953)
The `ClientBuilder` struct is gated behind the `gateway` feature, which
means that the client will receive the current application ID over the
gateway when first authenticating. Therefore it doesn't make sense to
set the application ID through the `ClientBuilder` as it will get
overrided anyway.

Also, the `get_token` method is pretty useless because the only source
of the token would be whatever the user would have passed to
`ClientBuilder::new`, meaning they would already have a way to read out
its value.